% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toxtools-package.R
\docType{package}
\name{toxtools-package}
\alias{toxtools}
\alias{toxtools-package}
\title{toxtools: Utilities for Ecotoxicological Data Handling and Analysis}
\description{
A collection of utility functions supporting ecotoxicological workflows, including data preparation, summarisation and helpers shared across concentration-response and species sensitivity distribution analyses.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://open-aims.github.io/toxtools/}
  \item \url{https://github.com/open-AIMS/toxtools}
  \item Report bugs at \url{https://github.com/open-AIMS/toxtools/issues}
}

}
\author{
\strong{Maintainer}: Rebecca Fisher \email{r.fisher@aims.gov.au}

Authors:
\itemize{
  \item Rebecca Fisher \email{r.fisher@aims.gov.au}
}

Other contributors:
\itemize{
  \item Australian Institute of Marine Science [copyright holder, funder]
}

}
\keyword{internal}
